requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
google-genai==1.0.0
langchain==0.3.13
langchain-core==0.3.28
streamlit==1.41.1
requests==2.32.3
pillow==11.0.0
